refactor: Update BusinessDayMapping to use Float64Array for improved precision
- Changed the type of `scaledX` and `timeByIndex` in the BusinessDayMapping interface from Float32Array to Float64Array to enhance numerical precision.
- Updated the `mapToBusinessDayScale` and `applyBusinessDayX` functions to accommodate the new Float64Array types, ensuring consistency across time scale calculations.

+++ src/core/time/applyTimeScale.ts
@@ -30,7 +30,7 @@ export function resolveTimeScaleOpts(xAxis: AxisOptions): TimeScaleOptions {
 export function applyBusinessDayX(
   x: Float32Array | Float64Array,
   xAxis: AxisOptions,
-): { displayX: Float32Array; mapping: BusinessDayMapping } {
+): { displayX: Float64Array; mapping: BusinessDayMapping } {
   const mapping = mapToBusinessDayScale(x, resolveTimeScaleOpts(xAxis));
   return { displayX: mapping.scaledX, mapping };
 }
